🚧 8Y24PK task: apply external agent result
