+++ .agentplane/tasks/202608170928-8Y24PK/pr/diffstat.txt
@@ -1,10 +1,11 @@
  docs/recipes/hermes-agentplane.mdx                 |  28 +++-
- docs/workflow-guides/hermes-kanban.mdx             |  94 ++++++-----
- integrations/hermes-agentplane-plugin/README.md    | 106 +++++++++---
+ docs/workflow-guides/hermes-kanban.mdx             |  94 ++++++------
+ integrations/hermes-agentplane-plugin/README.md    | 106 ++++++++++----
  .../lane-registry.example.json                     |  19 +++
  .../protocol-v2.schema.json                        |  39 +++++
- .../src/commands/hermes/hermes-environment.ts      |  65 +++++++-
- .../src/commands/hermes/hermes-runtime.ts          |  35 ++--
- .../src/commands/hermes/hermes.command.test.ts     | 181 ++++++++++++++++++++-
- .../src/commands/hermes/hermes.command.ts          |  52 +++++-
- 9 files changed, 519 insertions(+), 100 deletions(-)
+ .../src/commands/hermes/hermes-environment.test.ts |  30 ++++
+ .../src/commands/hermes/hermes-environment.ts      |  65 ++++++++-
+ .../src/commands/hermes/hermes-runtime.ts          |  35 +++--
+ .../src/commands/hermes/hermes.command.test.ts     | 157 ++++++++++++++++++++-
+ .../src/commands/hermes/hermes.command.ts          |  52 ++++++-
+ 10 files changed, 525 insertions(+), 100 deletions(-)
