+++ .agentplane/tasks/202608170928-8Y24PK/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202608170928-8Y24PK/upgrade-the-hermes-agentplane-bridge-protocol-ac",
   "created_at": "2026-08-17T09:29:36.876Z",
-  "diffstat_sha256": "sha256:0532d8a10bd6a24734b367ed2d9713bc92e3bab6ce606f511faa8240dcba9f57",
+  "diffstat_sha256": "sha256:8089bd8770dcec7b701b7609c7013a5ec3d35d5f36a8b44e8e5bb247c6ebdf43",
   "last_verified_at": "2026-08-17T23:02:48.265Z",
   "last_verified_diffstat_sha256": "sha256:0532d8a10bd6a24734b367ed2d9713bc92e3bab6ce606f511faa8240dcba9f57",
   "pr_number": 4837,
